Move temporary backward compatibility code to _path_finder/temporary_backward_compatibility.py, to fully sanitize path_finder.py
