+++ cuda_bindings/cuda/bindings/_path_finder/README.md
@@ -1,1 +1,3 @@
 # The `cuda.bindings.path_finder` module was moved → `cuda.pathfinder`
+
+`cuda.bindings.pathfinder` is slated to be removed after the next `cuda-bindings` release.
